% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePairwiseEstimates.R
\name{computePairwiseEstimates}
\alias{computePairwiseEstimates}
\title{Compute all pairwise estimates}
\usage{
computePairwiseEstimates(fit)
}
\arguments{
\item{fit}{the output of \code{fitNetwork}}
}
\description{
After fitting the model, this function pulls out all pairwise network effect (direct and indirect
combined) in a tidy format, on the hazard rate scale.
}
