% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRankProbabilities.R
\name{computeRankProbabilities}
\alias{computeRankProbabilities}
\title{Compute the rank probabilities}
\usage{
computeRankProbabilities(fit, preferredDirection = 1)
}
\arguments{
\item{fit}{object returned from GeMTC.}

\item{preferredDirection}{scalar, preferential direction of the outcome. \code{1} means higher
values are preferred, \code{-1} means lower values are preferred.}
}
\value{
Tidy data frame.
}
\description{
To be used for, i.a., rankograms (that use the cumulative probability).
}
