% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRhat.R
\name{computeRhat}
\alias{computeRhat}
\title{Compute the Rhat statistic for a given fit}
\usage{
computeRhat(fit)
}
\arguments{
\item{fit}{an object of class \code{"mtc.result"} or \code{"mtc.nodesplit"}.}
}
\value{
Tidy tibble. If a node-splitting result is supplied, the first columns called
\code{modelName} indicates the comparison that was split in the analysis.
}
\description{
Returns a tidy tibble with the Rhat statistic for each parameter. Also handles node-splitting
models.
}
