% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSucra.R
\name{computeSucra}
\alias{computeSucra}
\title{Compute SUCRA estimates}
\usage{
computeSucra(rankProbabilities)
}
\arguments{
\item{rankProbabilities}{output of \code{computeRankProbabilities}}
}
\description{
Computes surface under the cumulative ranking curve (SUCRA) for each treatment in the network.
The ranking curve and rankogram are synonyms. DISLCAIMER: Not sure this deserves its own
function. Let's see down the road.
}
