% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveEdges.R
\name{deriveEdges}
\alias{deriveEdges}
\title{Use to prepare data needed to draw the edges of the network}
\usage{
deriveEdges(networkData, nodes)
}
\arguments{
\item{networkData}{subset of the original network study data.}

\item{nodes}{output from \code{deriveNodes}}
}
\value{
A ggplot object.
}
\description{
Use to prepare data needed to draw the edges of the network
}
