% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriveNodes.R
\name{deriveNodes}
\alias{deriveNodes}
\title{Deriving node data for drawing the network}
\usage{
deriveNodes(networkData, labelOrientation)
}
\arguments{
\item{labelOrientation}{string giving how to rotate the label. There are two allowed settings:
\code{"radial"} (default) and \code{"horisontal"}/\code{"horizontal"}.}
}
\value{
A tidy data frame.
}
\description{
Deriving node data for drawing the network
}
