% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchParameterSummaries.R
\name{fetchParameterSummaries}
\alias{fetchParameterSummaries}
\title{Table of parameter estimates and deviance statistics for reconciled results}
\usage{
fetchParameterSummaries(
  parameterSummaries,
  devianceStatistics,
  networkMetaAnalysisId = NULL,
  nDigits = 1
)
}
\arguments{
\item{parameterSummaries}{output of \code{deriveParameterQuantiles}.}

\item{devianceStatistics}{output of \code{fetchDevianceStatistics}.}

\item{networkMetaAnalysisId}{scalar, what analysis to use. If \code{NULL} (default), the function
assumes pre-filtered data to be supplied in the first two arguments.}

\item{nDigits}{scalar, how many digits to show for parameter estimates. Won't affect deviance
statistics.}
}
\description{
Table of parameter estimates and deviance statistics for reconciled results
}
