% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPosteriorDraws.R
\name{fetchPosteriorDraws}
\alias{fetchPosteriorDraws}
\title{Fetches and tidies the posterior draws}
\usage{
fetchPosteriorDraws(fit, drawsThin = 10)
}
\arguments{
\item{fit}{the result of \code{fitNetwork}.}

\item{drawsThin}{scalar, thinning factor. The default value (\code{10}) keeps every tenth observation.
This is mainly used when producing output to be saved to the server, to keep the storage
requirements at a reasonable level.}
}
\value{
Tidy \code{tibble} with columns for chain and MCMC iteration indicators and one column
per parameter.
}
\description{
Potentially, with some thinning to keep the amount of space required down.
}
