% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTcoEstimates.R
\name{fetchTcoEstimates}
\alias{fetchTcoEstimates}
\title{Load target-comparator-outcome estimates from server}
\usage{
fetchTcoEstimates(conn, resultsCdm, resultsTable)
}
\arguments{
\item{conn}{a connection to a database, e.g., the result of calling
\code{DatabaseConnector::connect}.}

\item{resultsCdm}{string, the name of the schema in which the table lives.}

\item{resultsTable}{string, the name of the table holding the estimates.}
}
\description{
Load TCO estimates from the server. They need some processing (\code{wrangleTcoEstimates}) before
they are ready.
}
