% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchTidyNodesplitResults.R
\name{fetchTidyNodesplitResults}
\alias{fetchTidyNodesplitResults}
\title{Get the node-splitting analyses results in tidy format}
\usage{
fetchTidyNodesplitResults(fit)
}
\arguments{
\item{fit}{the result of calling \code{fitNetwork} with \code{includeNodesplittingAnalysis = TRUE}.}
}
\value{
Tidy data frame.
}
\description{
Returns a tidy data frame with direct, indirect and consistency (network) effect estimates for each
target-comparator pair along with the probability of inconsistency between the direct and
indirect estimates.
}
