% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNetwork.R
\name{fitNetwork}
\alias{fitNetwork}
\title{Run model on network to estimate parameters}
\usage{
fitNetwork(
  networkData,
  includeNodesplittingAnalysis = FALSE,
  modelType = "random",
  nWarmup = 1000,
  nIter = 3000,
  fitThin = 1,
  nChains = 4
)
}
\arguments{
\item{networkData}{a \code{dataframe} holding the subset of a target-comparator-outcome estimates
from the result of \code{fetchTcoEstimates}, with specified \code{outcomeId} and
\code{analysisId}.}

\item{includeNodesplittingAnalysis}{boolean, should also node-splitting analyses be run? NB: This
can be quite time-consuming, so is by default set to \code{FALSE}.}

\item{modelType}{string defining whether to run a random-effects (\code{"random"}, default) or a
fixed-effects (\code{"fixed"}) mode.}

\item{nWarmup}{scalar, the number of iterations the sampler should do before it starts sampling
from the posterior. This should be sufficient to the chains to have converged.}

\item{nIter}{scalar, the number of posterior samples to draw per chain.}

\item{fitThin}{scalar, thinning factor. \code{1} (default) keeps all samples whereas e.g.
\code{10} would keep only every tenth sample.}

\item{nChains}{scalar, the number of JAGS chains to run}
}
\description{
Runs a GeMTC-powered network meta-analysis on a subset of the results of calling
\code{fetchTcoEstimates} (i.e., setting analysisId and outcomeId) and
}
