% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShinyApp.R
\name{launchShinyApp}
\alias{launchShinyApp}
\title{Launch a Shiny app to explore the results}
\usage{
launchShinyApp(results, rstudio = TRUE, ...)
}
\arguments{
\item{results}{the output of \code{reconcileResults} directly or \code{loadFromDb} if the results
live on a server.}

\item{rstudio}{boolean, should the app run in the Viewer window (\code{TRUE}, default) or in an
internet browser (\code{FALSE})?}

\item{...}{passed on to \code{shiny::runApp}}
}
\description{
Launch a Shiny app to explore the results
}
