% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromDatabase.R
\name{loadFromDatabase}
\alias{loadFromDatabase}
\title{Loads results stored in a database following its data model}
\usage{
loadFromDatabase(conn, schema)
}
\arguments{
\item{conn}{connection to database, e.g. result of calling \code{DatabaseConnector::connect}}

\item{schema}{string, the name of the schema where results tables live.}
}
\description{
Loads results stored in a database following its data model
}
