% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotForest.R
\name{plotForest}
\alias{plotForest}
\title{Draw forest plot of total network estimates}
\usage{
plotForest(
  relativeEffectsEstimates,
  reference = NULL,
  lineSize = 0.5,
  pointSize = 1
)
}
\arguments{
\item{relativeEffectsEstimates}{the output of \code{computeRelativeEffects}.}

\item{reference}{string, which exposure to use as the reference. Choice is arbitrary. If nothing
is supplied, the first value in the data frame is used.}

\item{lineSize, pointSize}{scalars}
}
\description{
Using a \code{fitNetwork} object, this functions draws a classic forest plot of the combined
(direct + indirect) effects estimated.
}
