% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Visualise network}
\usage{
plotNetwork(
  edges,
  nodes,
  maxNodeRadius = 0.1,
  nodeColour = "dodgerblue",
  labelOrientation = "radial",
  labelColour = "black",
  edgeColour = "dodgerblue",
  edgeAlpha = 0.2
)
}
\arguments{
\item{maxNodeRadius}{the radius of the largest node in the network on the original
coordinate scales (= [-1, [1). Default = 0.1.}

\item{labelOrientation}{string giving how to rotate the label. There are two allowed settings:
\code{"radial"} (default) and \code{"horisontal"}/\code{"horizontal"}.}

\item{edgeColour, nodeColour, labelColour}{strings given the colours of edges, nodes and label
text.}

\item{edgeAlpha}{scalar in [0, 1] giving the opacity of the edges (default: 0.2). 1 = complete
opague, 0 = completely transparent.}

\item{treatmentMetaData}{NOT IN USE a \code{dataframe} with information about each node in the
network, e.g., name and databases with data for this treatment.}

\item{edgeMetaData}{NOT IN USE a \code{dataframe} with information about the head-to-head
comparisons. Not that the number of head-to-head comparisons is computed so doesn't need to be
supplied.}
}
\description{
Draw the head-to-head comparisons as a network. The function relies on \pkg{ggraph} to do the
heavy lifting.
}
