% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNodesplitResults.R
\name{plotNodesplitResults}
\alias{plotNodesplitResults}
\title{Visualise results of node-splitting analysis}
\usage{
plotNodesplitResults(nodesplitResults)
}
\arguments{
\item{nodesplitResults}{tidy data frame with consistency, direct and indirect HR estimates with CrIs.}
}
\description{
Pending, but could be useful for getting an overview.
}
