% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairwiseEstimates.R
\name{plotPairwiseEstimates}
\alias{plotPairwiseEstimates}
\title{Creates a heatmap plot of all pairwise estimates}
\usage{
plotPairwiseEstimates(
  tidyEstimates,
  nDigits = 1,
  colour_low = "blue",
  colour_high = "red",
  contrastThreshold = 0.5,
  mid = "hr",
  lo = "cri95Lb",
  hi = "cri95Ub",
  textSize = 12,
  sucraEstimates = NULL
)
}
\arguments{
\item{tidyEstimates}{data frame/tibble with all pairwise combinations to feature in the heatmap.}

\item{nDigits}{scalar, how many digits should the shown estimates have? Default is \code{1}}

\item{colour_low, colour_high}{valid colour specifications for the lower and higher bounds of the
diverging colour scale}

\item{contrastThreshold}{scalar, absolute values on the log-scale above this threshold will be
white, otherwise they will be black}

\item{mid, lo, hi}{strings, the names of the columns in the \code{tidyEstimates} data frame holding
the point estimate (usually median) and lower and upper bounds of the credibility intervals
(usually 2.5 and 97.5 percentiles)}

\item{sucraEstimates}{data frame with two columns: exposureId and SUCRA estimate as numeric
value. Output from \code{computeSucra}. Default: \code{NULL}. If valid data supplied, the SUCRA
estimates will be shown in the diagonal, and the exposures will be ordered by descending SUCRA
estimate.}
}
\description{
Some customisation possible, but this is still pretty basic.
}
