% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRankogram.R
\name{plotRankogram}
\alias{plotRankogram}
\title{Rank treatments by effectiveness}
\usage{
plotRankogram(rankProbabilities, lineSize = 1, lineAlpha = 0.9)
}
\arguments{
\item{rankProbabilities}{output of \code{computeRankProbabilities}}

\item{lineSize}{scalar, thickness of the plot lines.}

\item{lineAlpha}{the opacity of drawn lines. If \code{alpha = 1}, lines will be entirely opaque
masking overlain lines, so the default is \code{0.7}.}
}
\value{
A ggplot object, or a list of ggplot objects.
}
\description{
Produces a \pkg{ggplot2} object so can be further customised as desired. If the results of a
node-splitting analysis is supplied, the output is a list of plots, one for each analysis and one
for the consistency fit.
}
