% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRhat.R
\name{plotRhat}
\alias{plotRhat}
\title{Plot the Rhat values of the fitted model}
\usage{
plotRhat(
  rhat,
  threshold = 1.05,
  showParameters = FALSE,
  pointSize = 1,
  lineSize = 0.5
)
}
\arguments{
\item{rhat}{output from \code{computeRhat}}

\item{threshold}{scalar, what is the threshold below which you want all Rhat values to lie?
Default is \code{1.05}. Set to \code{NULL} to remove the indicator in the plot.}

\item{showParameters}{boolean, should parameters be distinguishable with colours? Potentially
useful if any are above the threshold, so by default set to \code{FALSE}.}

\item{pointSize}{scalar, size of points representing Rhat values.}

\item{lineSize}{scalar, thickness of vertical indicator of threshold, if shown.}
}
\value{
A \code{ggplot} object.
}
\description{
Instead of looking at the Rhat values in table format, it's usually easier to quickly scan a
plot, and this function allows you to do just that.
}
