% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraces.R
\name{plotTraces}
\alias{plotTraces}
\title{Produces density plots of parameter estimates from the fitted models}
\usage{
plotTraces(
  posteriorDraws,
  separateChains = TRUE,
  lineAlpha = 0.7,
  lineSize = 0.2,
  wrapByModelName = FALSE
)
}
\arguments{
\item{separateChains}{boolean, should chains be plotted as separate distributions (\code{TRUE},
default) or as one (\code{FALSE})? Not plotting chains separately somewhat defeats the purpose
of the visualisation but is anyway left to the user.}

\item{lineAlpha}{the opacity of drawn lines. If \code{alpha = 1}, lines will be entirely opague
masking overlain lines, so the default is \code{0.7}.}

\item{lineSize}{scalar, thickness of the plot lines.}

\item{modelFit}{an object containing the fitted results.}
}
\value{
A ggplot object, or a list of ggplot objects.
}
\description{
Produces a \pkg{ggplot2} object so can be further customised as desired. If the results of a
node-splitting analysis is supplied, the output is a list of plots, one for each analysis and one
for the consistency fit.
}
