% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileResults.R
\name{reconcileResults}
\alias{reconcileResults}
\title{Puts the results together}
\usage{
reconcileResults(networkMetaAnalysisResults, discardPosteriorDraws = FALSE)
}
\arguments{
\item{networkMetaAnalysisResults}{the output of \code{runAnalyses}.}

\item{discardPosteriorDraws}{boolean, should posterior draws be kept? These can take up quite
some space, so there might be use cases in which one doesn't want to keep them. They are
useful, however, for sampling diagnostics so are kept by default.}
}
\description{
Gives the data a format appropriate for the database, but could also be used in R directly.
}
