% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleNetworkMetaAnalysis.R
\name{runSingleNetworkMetaAnalysis}
\alias{runSingleNetworkMetaAnalysis}
\title{Run a single Network Meta Analysis}
\usage{
runSingleNetworkMetaAnalysis(
  nmaId,
  aggregatedResults,
  includeNodesplittingAnalysis = FALSE,
  nWarmup = 1000,
  nIter = 3000,
  nChains = 4,
  fitThin = 1,
  referenceTreatmentId = NULL,
  preferredDirection = 1,
  modelType = "random",
  drawsThin = 10
)
}
\arguments{
\item{aggregatedResults}{tidy data frame, result of calling \code{fetchTcoEstimates}.}

\item{includeNodesplittingAnalysis}{boolean, should also node-splitting analyses be run? NB: This
can be quite time-consuming, so is by default set to \code{FALSE}.}

\item{nWarmup}{scalar, the number of iterations the sampler should do before it starts sampling
from the posterior. This should be sufficient to the chains to have converged.}

\item{nIter}{scalar, the number of posterior samples to draw per chain.}

\item{nChains}{scalar, the number of JAGS chains to run}

\item{fitThin}{scalar, thinning factor. \code{1} (default) keeps all samples whereas e.g.
\code{10} would keep only every tenth sample.}

\item{referenceTreatmentId}{string, the name of the exposure to use as the reference when
computing pairwise estimates. If none given, the first in alphabetical order will be used.}

\item{preferredDirection}{scalar, preferential direction of the outcome. \code{1} means higher
values are preferred, \code{-1} means lower values are preferred.}

\item{modelType}{string defining whether to run a random-effects (\code{"random"}, default) or a
fixed-effects (\code{"fixed"}) mode.}

\item{drawsThin}{scalar, thinning factor used when fetching posterior draws to keep in the
database. Used to produce further derived quantities, density plots of posterior distributions
and trace plots.}
}
\description{
Workhorse of the analysis pipeline. This is an internal function invoked either for sequential
(\code{nCores = 1}) or parallel processing (\code{nCores > 1}). Probably shouldn't be called by
the user.
}
