% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToDatabase.R
\name{saveToDatabase}
\alias{saveToDatabase}
\title{Saves the reconciled results to database}
\usage{
saveToDatabase(
  conn,
  reconciledResults,
  schema,
  overwriteExistingTables = FALSE
)
}
\arguments{
\item{conn}{connection object.}

\item{reconciledResults}{output of \code{reconcileResults}.}

\item{schema}{string, the name of the schema where the tables will live. Should probably be a
dedicated schema to prevent conflicting names.}

\item{overwriteExistingTables}{boolean, should tables be overridden? This is a bit dangerous and
so set to \code{FALSE} by default. See note on dedicated schema above.}
}
\description{
If desired, this function takes the reconciled results and saves them to the database for
downstream use.
}
