% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleTcoEstimates.R
\name{wrangleTcoEstimates}
\alias{wrangleTcoEstimates}
\title{Putting original data into appropriate format}
\usage{
wrangleTcoEstimates(tcoEstimates, excludedDatabases = "Meta-analysis")
}
\arguments{
\item{tcoEstimates}{data frame/tibble, in the original LEGEND format.}

\item{excludedDatabases}{character vector, names of the databases whose results should be
ignored. The default (\code{"Meta-analysis"}) is obvious and should be kept if other databases
are added.}
}
\description{
Whether data were loaded from a database o a file, they need some wrangling to play well with the
rest of the workflow in the package. This function does just that.
}
