% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNodesplitResults.R
\name{plotNodesplitResults}
\alias{plotNodesplitResults}
\title{Visualise results of node-splitting analysis}
\usage{
plotNodesplitResults(nodesplitResults)
}
\description{
Pending, but will be useful for getting an overview.
}
