% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraces.R
\name{plotTraces}
\alias{plotTraces}
\title{Produces density plots of parameter estimates from the fitted models}
\usage{
plotTraces(
  posteriorDraws,
  separateChains = TRUE,
  lineAlpha = 0.7,
  lineSize = 0.2,
  wrapByModelName = FALSE
)
}
\arguments{
\item{posteriorDraws}{tidy data frame, all posterior draws with (at least) the following columns:
\code{modelName}, \code{chainId}, \code{mcmcIterationId}, \code{parameterName},
\code{parameterValue}.}

\item{separateChains}{boolean, should chains be plotted as separate distributions (\code{TRUE},
default) or as one (\code{FALSE})? Not plotting chains separately somewhat defeats the purpose
of the visualisation but is anyway left to the user.}

\item{lineAlpha}{the opacity of drawn lines. If \code{alpha = 1}, lines will be entirely opague
masking overlain lines, so the default is \code{0.7}.}

\item{lineSize}{scalar, thickness of the plot lines.}

\item{wrapByModelName}{boolean, if the \code{modelName} column contains more than one distinct
value, this should be set to \code{TRUE} to show separate plots for each model. Otherwise, the
plot will combine samples for the same parameter from different models, which will give
misleading results.}
}
\value{
A ggplot object, or a list of ggplot objects.
}
\description{
Produces a \pkg{ggplot2} object so can be further customised as desired. If the results of a
node-splitting analysis is supplied, the output is a list of plots, one for each analysis and one
for the consistency fit.
}
