% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleTcoEstimates.R
\name{wrangleTcoEstimates}
\alias{wrangleTcoEstimates}
\title{Putting original data into appropriate format}
\usage{
wrangleTcoEstimates(tcoEstimates)
}
\arguments{
\item{tcoEstimates}{data frame/tibble, in the original LEGEND format.}
}
\description{
Whether data were loaded from a database o a file, they need some wrangling to play well with the
rest of the workflow in the package. This function does just that.
}
